+++ src/assets/empty/empty-folder.tsx
@@ -1,0 +1,53 @@
+import React from 'react';
+
+export const emptyFolder = (
+    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="41 42 124 106">
+        <path
+            className="fill-100"
+            d="M150.941 132.426V66.624a4 4 0 0 0-4-4H99.969a4 4 0 0 1-4-4V57.1a4 4 0 0 0-4-4H44.997a4 4 0 0 0-4 4v75.325a4 4 0 0 0 4 4h101.944a4 4 0 0 0 4-4Z"
+        />
+        <g filter="url(#filter-folder)">
+            <path
+                className="fill-white"
+                d="M71.978 144V91.68a4 4 0 0 1 4-4h35.507a3.892 3.892 0 0 0 3.892-3.891 3.891 3.891 0 0 1 3.891-3.892h35.507a4 4 0 0 1 4 4V144a4 4 0 0 1-4 4H75.978a4 4 0 0 1-4-4Z"
+            />
+            <path
+                className="stroke-400"
+                d="M72.478 91.68V144a3.5 3.5 0 0 0 3.5 3.5h78.797a3.5 3.5 0 0 0 3.5-3.5V83.897a3.5 3.5 0 0 0-3.5-3.5h-35.507a3.391 3.391 0 0 0-3.391 3.392 4.392 4.392 0 0 1-4.392 4.392H75.978a3.5 3.5 0 0 0-3.5 3.5Z"
+            />
+        </g>
+        <path
+            className="fill-200"
+            d="M55 108a4 4 0 0 1 4-4h34a4 4 0 0 1 0 8H59a4 4 0 0 1-4-4ZM55 121a4 4 0 0 1 4-4h66a4 4 0 0 1 0 8H59a4 4 0 0 1-4-4Z"
+        />
+        <path
+            className="fill-400"
+            d="M165 82a8 8 0 1 0-16 0 8 8 0 0 0 16 0ZM71.996 54.571a2.67 2.67 0 0 0-3.777 0l-11.33 11.33a2.67 2.67 0 0 0 3.776 3.778l11.33-11.331a2.67 2.67 0 0 0 0-3.777Z"
+        />
+        <path
+            className="fill-white stroke-400"
+            stroke-width="2"
+            d="M74.113 43c-5.347 0-9.683 4.335-9.683 9.683 0 5.347 4.335 9.683 9.683 9.683s9.683-4.336 9.683-9.683c0-5.348-4.335-9.683-9.683-9.683Z"
+        />
+        <defs>
+            <filter
+                id="filter-folder"
+                width="102.798"
+                height="84.103"
+                x="63.978"
+                y="74.897"
+                color-interpolation-filters="sRGB"
+                filterUnits="userSpaceOnUse"
+            >
+                <feFlood flood-opacity="0" result="BackgroundImageFix" />
+                <feColorMatrix in="SourceAlpha" result="hardAlpha" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" />
+                <feOffset dy="3" />
+                <feGaussianBlur stdDeviation="4" />
+                <feComposite in2="hardAlpha" operator="out" />
+                <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.1 0" />
+                <feBlend in2="BackgroundImageFix" result="effect1_dropShadow_5857_64030" />
+                <feBlend in="SourceGraphic" in2="effect1_dropShadow_5857_64030" result="shape" />
+            </filter>
+        </defs>
+    </svg>
+);