+++ stories/02-Design System/05-DataDisplay/Empty/examples/basic.tsx
@@ -1,0 +1,8 @@
+import {KitEmpty} from '@kit/DataDisplay';
+import React from 'react';
+
+const App = () => {
+    return <KitEmpty />;
+};
+
+export default App;